requirements.txt
# Core Dependencies
ultralytics==8.3.252
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
Pillow==10.2.0
numpy==1.26.3

# Machine Learning & AI
torch==2.1.2
torchvision>=0.16.0
onnx>=1.14.0
onnxruntime>=1.15.0

# LLM Integration
openai>=1.10.0
google-generativeai>=0.3.0
anthropic>=0.8.0
langchain>=0.1.0
langchain-openai>=0.0.5

# Web Framework
streamlit==1.52.1
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database & Caching
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pymysql>=1.1.0
alembic>=1.13.0
redis>=5.0.0
pymongo>=4.6.0

# File Processing
reportlab>=4.0.0
pypdf2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
python-pptx>=0.6.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
requests>=2.31.0
httpx>=0.26.0
aiohttp>=3.9.0
aiofiles>=23.2.0
onvif-zeep>=0.2.12

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0
sentry-sdk==1.39.2

# Data Validation
marshmallow==3.20.2
cerberus==1.3.5

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.3
pytest-mock==3.12.0
pytest-xdist==3.5.0
faker==22.0.0
responses==0.24.1
hypothesis>=6.90.0

# Code Quality
black==24.1.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0
pylint==3.0.3
bandit==1.7.6

# Development Tools
ipython==8.20.0
jupyter==1.0.0
notebook==7.0.7
pre-commit==3.6.0

# Deployment
gunicorn==21.2.0
docker==7.0.0
boto3==1.34.25
google-cloud-storage==2.14.0

# CLI Tools
click==8.1.7
typer==0.9.0
rich==13.7.0
tqdm==4.66.1

# Security
cryptography==42.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
msal>=1.24.0

# Rate Limiting & Caching
slowapi==0.1.9
cachetools==5.3.2

# Image Processing
imageio==2.33.1
scikit-image==0.22.0

# Async Processing
celery==5.3.6
kombu==5.3.5

# API Documentation
mkdocs==1.5.3
mkdocs-material==9.5.6

# Time & Date
python-dateutil==2.8.2
pytz==2023.3.post1

# Data Visualization
plotly==5.18.0